divvy-shim-openwhisk/shim/requirements.txt
wiotp-sdk==0.9.5
requests==2.22.0